Add interface methods for custody transfers
