+++ client/client.go
@@ -1203,3 +1203,38 @@ func (bc *Blockchain) RewarderTotalClaimed(rewarderAddress common.Address) (*big
 		Context: ctx,
 	})
 }
+
+type CustodyTokensTransfer struct {
+	WriteRequest
+	CustodyAddress common.Address
+	Recipients     []common.Address
+	Amounts        []*big.Int
+}
+
+func (bc *Blockchain) CustodyTransferTokens(req CustodyTokensTransfer) (*types.Transaction, error) {
+	transactor, err := rewarder.NewCustodyTransactor(req.CustodyAddress, bc.ethClient.Client())
+	if err != nil {
+		return nil, err
+	}
+
+	ctx, cancel := context.WithTimeout(context.Background(), bc.bcTimeout)
+	defer cancel()
+
+	nonce, err := bc.getNonce(req.Identity)
+	if err != nil {
+		return nil, errors.Wrap(err, "could not get nonce")
+	}
+
+	return transactor.Payout(
+		&bind.TransactOpts{
+			From:     req.Identity,
+			Signer:   req.Signer,
+			Context:  ctx,
+			GasLimit: req.GasLimit,
+			GasPrice: req.GasPrice,
+			Nonce:    big.NewInt(0).SetUint64(nonce),
+		},
+		req.Recipients,
+		req.Amounts,
+	)
+}
